Disable KCAL Inversion toggle , it's unimportant and bugged on AICP
It was just a boring bug, and the "Color Inversion" toggle in the "Accessibility" section in the "Settings" app still works flawlessly, so if user needs color inversion will use that option (which even has a QS tile).
I don't exactly know why it's bugged on AICP, logcats and dmesg and event logs AFAIK are the same as the MiracleDroid Oreo ones where it works flawlessly.